Fixed bug where triggerFetchAndRedraw() was being called on both create() and createView() of PumpListFragment, causing all pumps to be fetched twice.
Straightened the offline storage policy a bit on first load of app.
XXX: There's still a bug as it can't be determined if a set of objects is already pinned. Need to follow up.

Used query.fromPin() when a pump is queried for based on its ID in subsequent activities.